requirements.txt
fire
sentencepiece
torch>=2.2.0
xformers>=0.0.22
tiktoken
blobfile
flask
flask_cor
einops
transformers